requirements.txt
CherryPy==11.0.0
pytest==3.2.3
google-cloud